+++ ios/engine/KMEI/KeymanEngine/Classes/Constants.swift
@@ -50,6 +50,11 @@ public enum Key {
 
   // Other keys
   static let update = "update"
+  
+  // ResourceDownloadQueue keys
+  static let downloadTask = "downloadTask"
+  static let downloadBatch = "downloadBatch"
+  static let downloadQueueFrame = "queueFrame"
 }
 
 public enum Defaults {