+++ ios/engine/KMEI/KeymanEngine/Classes/LanguagePicker/LanguageDetailViewController.swift
@@ -130,7 +130,7 @@ class LanguageDetailViewController: UITableViewController, UIAlertViewDelegate {
       cell.detailTextLabel?.isEnabled = true
     }
 
-    let kbState = Manager.shared.stateForKeyboard(withID: keyboard.id)
+    let kbState = ResourceDownloadManager.shared.stateForKeyboard(withID: keyboard.id)
     cell.setKeyboardState(kbState, selected: false, defaultAccessoryType: cell.accessoryType)
   }
 
@@ -139,7 +139,7 @@ class LanguageDetailViewController: UITableViewController, UIAlertViewDelegate {
     let keyboardIndex = indexPath.section
     let keyboard = language.keyboards![keyboardIndex]
 
-    let state = Manager.shared.stateForKeyboard(withID: keyboard.id)
+    let state = ResourceDownloadManager.shared.stateForKeyboard(withID: keyboard.id)
     if state != .downloading {
       if state == .needsDownload {
         isUpdate = false
@@ -162,7 +162,7 @@ class LanguageDetailViewController: UITableViewController, UIAlertViewDelegate {
   }
 
   func downloadHandler(_ keyboardIndex: Int) {
-    Manager.shared.downloadKeyboard(withID: language.keyboards![keyboardIndex].id,
+    ResourceDownloadManager.shared.downloadKeyboard(withID: language.keyboards![keyboardIndex].id,
                                       languageID: language.id, isUpdate: isUpdate)
   }