+++ ios/keyman/Keyman/Keyman/GetStartedViewController/GetStartedViewController.swift
@@ -142,7 +142,7 @@ class GetStartedViewController: UIViewController, UITableViewDelegate, UITableVi
     switch indexPath.section {
     case 0:
       mainViewController.dismissGetStartedView(nil)
-      Manager.shared.showKeyboardPicker(in: mainViewController, shouldAddKeyboard: true)
+      mainViewController.showInstalledLanguages()
     case 1:
       mainViewController.dismissGetStartedView(nil)
       let setUpVC = SetUpViewController()